requirements.txt
Django==4.2
django-crispy-forms==2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0